requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
fastapi==0.81.0
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
pydantic==1.10.1; python_version >= '3.7'
python-dotenv==0.20.0
requests==2.28.1; python_version >= '3.7' and python_version < '4'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
uvicorn==0.18.3
vk-api==11.9.9
